boost-build: Add improvements to the full example.
This adds install and documentation targets from the librfidtl
project.

Note that there is still an issue with installing the documentation
via the install target.  Only the redirect file api.html is copied,
the api directory is not.  This is due to the fact that the api.html
file is the actual target and the install rules are target-based.
This should be remedied at some point, but for now this is acceptable.